### 2020-03-01 #2127 @whedon
[PRE REVIEW]: audiomate: A Python package for working with audio datasets
**Submitting author:** @ynop (<a href="http://orcid.org/0000-0003-0207-5711">Matthias Büchi</a>)
**Repository:** <a href="https://github.com/ynop/audiomate" target ="_blank">https://github.com/ynop/audiomate</a>
**Version:** v5.2.0
