### 2020-03-02 #2127 @ynop
From the given list I would suggest:
- faroit
- desilinguist
- expectopatronum
