### 2020-03-02 #2127 @desilinguist
I don’t think I have the bandwidth for this right now but I can recommend
some of my colleagues: @mulhod and @bndgyawali on GitHub. They work a lot
on speech/audio processing.

On Mon, Mar 2, 2020 at 13:49 Fabian-Robert Stöter <notifications@github.com>
wrote:

> @terrytangyuan <https://github.com/terrytangyuan> with great pleasure!
>
