### 2020-03-04 ynop/audiomate#101 @mulhod
Sox dependency not mentioned in installation instructions
I am reviewing this repository for JOSS and I realized that I didn't have `sox` installed before running the unit tests. Some failed and then I tried again after installing `sox` and all succeeded. I think it might warrant a mention in the installation instructions.
