### 2020-03-04 #2135 @mulhod
Created issue about sox dependency: https://github.com/ynop/audiomate/issues/101
