### 2020-03-04 ynop/audiomate#102 @mulhod
Cannot get tests to finish on Mac with Python 3.8
After setting up a development environment on a Mac with python 3.8 and running `python setup.py test`, the tests would start to hang on `tests/corpus/conversion/test_wav.py`. CPU would become a bit overloaded and the tests would not advance (nor fail). After a while (maybe 2 hours because I wasn't sure how long it should take), I shut it off and tried it again and it was the same issue. I tried downgrading to python 3.6 just as a hunch (there seemed to issues with `loky`/`multiprocessing`, which I could see when I force-killed the tests) and the tests worked afterwards. No hanging, no failures.

This is the type of output I saw when using python 3.8 after I started running the tests and prematurely killed them after they started to hang:
<pre>
(/Users/mmulholland/Documents/audiomate_dev/audiomateenv) ➜  audiomate git:(master) ✗ python setup.py test
running pytest
[...]
running build_ext
===================================================================== test session starts =====================================================================
platform darwin -- Python 3.8.2, pytest-5.2.3, py-1.8.1, pluggy-0.13.1
benchmark: 3.2.2 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_i
terations=100000)
rootdir: /Users/mmulholland/Documents/audiomate_dev/audiomate, inifile: setup.cfg, testpaths: tests
plugins: requests-mock-1.7.0, benchmark-3.2.2, cov-2.8.1
collected 865 items

tests/test_docs.py .                                                                                                                                    [  0%]
tests/annotations/test_label.py ..........................................                                                                              [...]                                                                                                       [ 16%]
tests/corpus/test_corpus.py ......................................                                                                                      [ 21%]
tests/corpus/conversion/test_base.py ....                                                                                                               [ 21%]
tests/corpus/conversion/test_wav.py ...^C/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/site-packages/pytest_benchmark/logger.py:44: PytestBenchmarkWarning: Not saving anything, no benchmarks have been run!
  warner(PytestBenchmarkWarning(text))

====================================================================== warnings summary =======================================================================
tests/test_docs.py::TestDocs::test_html_docs
  /Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/site-packages/sphinx_rtd_theme/search.html:21: RemovedInSphinx30Warning: To modify script_files in the theme is deprecated. Please insert a <script> tag directly in your theme instead.
    {% endblock %}

-- Docs: https://docs.pytest.org/en/latest/warnings.html
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! KeyboardInterrupt !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/threading.py:302: KeyboardInterrupt
(to show a full traceback on KeyboardInterrupt use --full-trace)
============================================================== 189 passed, 1 warnings in 18.37s ===============================================================
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-5p3583zf'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-qmn4d63z'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-rae16u7s'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-r89lahv7'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-5o3lkl2d'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-c_mxohe5'
Traceback (most recent call last):
  File "/Users/mmulholland/Documents/audiomate_dev/audiomateenv/lib/python3.8/multiprocessing/resource_tracker.py", line 201, in main
    cache[rtype].remove(name)
KeyError: '/loky-60142-1u633727'
(/Users/mmulholland/Documents/audiomate_dev/audiomateenv) ➜  audiomate git:(master) ✗
</pre>

I see 3.8 is not included in the list of Python versions in [`.travis.yml`](https://github.com/ynop/audiomate/blob/master/.travis.yml). However, it seems like 3.8 is supposed to be supported. I should mention here also that I did the same thing on Linux and it worked whether it was 3.6 or 3.8.
