### 2020-03-04 #2135 @mulhod
Created issue about python 3.8 issues: https://github.com/ynop/audiomate/issues/102
