### 2020-03-04 ynop/audiomate#101 @ynop
Thanks for reviewing and your feedback.
I'll add a note in the installation instructions.
