### 2020-03-04 44b8f78
Add note for sox in installation instructions (fixes #101)
