+++ README.md
@@ -72,6 +72,22 @@ Install the latest development version:
 pip install git+https://github.com/ynop/audiomate.git
 ```
 
+### Dependencies
+
+#### sox
+For parts of the functionality (e.g. audio format conversion) [sox](http://sox.sourceforge.net) is used. In order to use it, you have to install sox.
+
+```sh
+# macos
+brew install sox
+
+# with support for specific formats
+brew install sox --with-lame --with-flac --with-libvorbis
+
+# linux
+apt-get install sox
+```
+
 ## Development
 
 ### Prerequisites
