### 2020-03-07 cd3892a
Update pytest-benchmark to v3.2.3
