### 2020-03-07 b7e216a
Update scipy to v1.4.1
