### 2020-03-07 c5fbccb
Update Sphinx to v2.4.4
