+++ setup.py
@@ -19,7 +19,7 @@ REQUIRED = [
     'networkx == 2.4',
     'requests == 2.23.0',
     'intervaltree == 3.0.2',
-    'tqdm == 4.38.0',
+    'tqdm == 4.43.0',
     'sox == 1.3.7',
     'PGet == 0.5.0',
 ]
