### 2020-03-07 ynop/audiomate#102 @aahlenst
The problem is multiprocessing in combination with pytest and setuptools/pip. This makes multiple tests hang.

If we want to continue invoking the tests using `python setup.py test`, we have to skip the tests:

```python
@pytest.mark.skipif(get_start_method() != 'fork')
```

Invoking the test suite using `python -m pytest` passes without having to skip the tests which I'd prefer. The best idea is probably to get rid of setuptools and use [poetry](https://python-poetry.org).
