### 2020-03-07 dd005cb
Recommend running test suite via pytest
Fixes #102.
