+++ .travis.yml
@@ -3,6 +3,7 @@ python:
     - "3.5"
     - "3.6"
     - "3.7"
+    - "3.8"
 
 install:
     - sudo apt-get update
