### 2020-03-10 ynop/audiomate#109 @mulhod
Failure during test_split_by_length_of_utterances benchmark test
When running `pytest bench`, I get a failure during the `test_split_by_length_of_utterances` test. See below. All other benchmark tests are successful.

I am using Python3.8.2 on OSX and Linux and I get the failure in the same benchmark test in both environments.

==================================================================== FAILURES ====================================================================
_______________________________________________________ test_split_by_length_of_utterances _______________________________________________________

benchmark = <pytest_benchmark.fixture.BenchmarkFixture object at 0x139d59670>

    def test_split_by_length_of_utterances(benchmark):
        corpus = resources.generate_corpus(
            179,
            (250, 500),
            (1, 9),
            (0, 6),
            (1, 20),
            random.Random(x=234)
        )

        splitter = subset.Splitter(corpus, random_seed=324)

>       benchmark(run, splitter)

bench/corpus/subset/splitting/test_split_by_length_of_utterances.py:33:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../audiomateenv/lib/python3.8/site-packages/pytest_benchmark/fixture.py:126: in __call__
    return self._raw(function_to_benchmark, *args, **kwargs)
../audiomateenv/lib/python3.8/site-packages/pytest_benchmark/fixture.py:148: in _raw
    duration, iterations, loops_range = self._calibrate_timer(runner)
../audiomateenv/lib/python3.8/site-packages/pytest_benchmark/fixture.py:276: in _calibrate_timer
    duration = runner(loops_range)
../audiomateenv/lib/python3.8/site-packages/pytest_benchmark/fixture.py:91: in runner
    function_to_benchmark(*args, **kwargs)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

splitter = <audiomate.corpus.subset.splitting.Splitter object at 0x139d595b0>

    def run(splitter):
>       out = splitter.split_by_length_of_utterances(
            {
                'train': 0.7,
                'test': 0.15,
                'dev': 0.15
            },
            separate_issuers=True
        )
E       AttributeError: 'Splitter' object has no attribute 'split_by_length_of_utterances'

bench/corpus/subset/splitting/test_split_by_length_of_utterances.py:9: AttributeError
