### 2020-03-10 #2135 @mulhod
Create issue about benchmark test failure: https://github.com/ynop/audiomate/issues/109
