### 2020-03-11 ynop/audiomate#109 @aahlenst
@ynop I guess it would make sense to run `pytest bench` as part of CI to prevent bit rot in the future? The run time seems moderate (less than 10 mins).
