### 2020-03-11 824f8a7
Fix splitter benchmark
Fixes #109.
