### 2020-03-11 83bf94c
Fix splitter benchmark
Fixes #109.
