### 2020-03-29 9ecc19e
Add progress messages.
Replace double quotes.

Update to new logging infrastructure.
