### 2020-03-29 0293548
Add progress messages.
Replace double quotes.

Update to new logging infrastructure.
