+++ audiomate/corpus/io/tatoeba.py
@@ -24,7 +24,8 @@ class TatoebaDownloader(base.CorpusDownloader):
           Website
 
     Args:
-        include_languages (list): List of languages to download. If None all are downloaded.
+        include_languages (list): List of languages codes to download. If None all are downloaded.
+                                  Languages: https://github.com/Tatoeba/tatoeba2/blob/dev/src/Lib/LanguagesLib.php#L222
         include_licenses (list): Sentences are downloaded only if their license is in this list.
                                  If None all licenses are included.
         load_empty_license (bool): Sentences with an empty license are not meant to be reused.
