### 2020-03-30 aac680c
Setup logging
(fixes #114)
