+++ audiomate/corpus/io/nvidia_jasper.py
@@ -2,13 +2,15 @@ import os
 import json
 import multiprocessing
 
-from tqdm import tqdm
 import librosa
 
 import audiomate
 from . import base
 
 from audiomate.corpus import conversion
+from audiomate import logutil
+
+logger = logutil.getLogger()
 
 
 class NvidiaJasperWriter(base.CorpusWriter):
@@ -82,11 +84,11 @@ class NvidiaJasperWriter(base.CorpusWriter):
 
         utts = []
 
-        print('Get utterance durations')
+        logger.info('Get utterance durations')
         utt_idx_to_path = {utt.idx: utt.track.path for utt in out_corpus.utterances.values()}
         utt_to_duration = self._get_file_durations(utt_idx_to_path)
 
-        print('Prepare utterance infos')
+        logger.info('Prepare utterance infos')
         for utterance in out_corpus.utterances.values():
             utt_dur = utt_to_duration[utterance.idx]
             utt_info = self.process_utterance(utterance, utt_dur, rel_base_path)
@@ -94,7 +96,7 @@ class NvidiaJasperWriter(base.CorpusWriter):
 
         utts = sorted(utts, key=lambda x: x[1]['original_duration'])
 
-        print('Prepare subviews')
+        logger.info('Prepare subviews')
         subviews = {}
         subviews['all'] = [u[1] for u in utts]
 
@@ -103,7 +105,7 @@ class NvidiaJasperWriter(base.CorpusWriter):
             subview_utts = [u[1] for u in utts if u[0] in utt_filter]
             subviews[subview_name] = subview_utts
 
-        print('Write files')
+        logger.info('Write files')
         for name, data in subviews.items():
             target_path = os.path.join(path, '{}.json'.format(name))
             with open(target_path, 'w') as f:
@@ -133,7 +135,10 @@ class NvidiaJasperWriter(base.CorpusWriter):
         num_files = len(file_items)
 
         with multiprocessing.Pool(self.num_workers) as p:
-            res = list(tqdm(p.imap(self._get_file_duration, file_items), total=num_files))
+            res = list(logger.progress(
+                p.imap(self._get_file_duration, file_items),
+                total=num_files
+            ))
 
         return {k: v for (k, v) in res}
 
