+++ audiomate/__init__.py
@@ -1,3 +1,4 @@
 from .corpus import Corpus  # noqa: F401
+from . import logutil  # noqa: F401
 
 __version__ = '5.2.0'
