### 2020-04-02 a3f5e7a
Use correct logger
Fixes #118
