### 2020-04-03 e28811a
Use logging instead of print/tqdm in downloader
fixes (#117)
