+++ audiomate/utils/download.py
@@ -8,7 +8,12 @@ import multiprocessing
 
 import requests
 from pget.down import Downloader
-from tqdm import tqdm
+
+from audiomate import logutil
+
+logger = logutil.getLogger()
+
+PROGRESS_LOGGER_BYTE_DELAY = 1024 * 1024 * 100
 
 
 def download_files(url_to_target, num_threads=1):
@@ -23,10 +28,10 @@ def download_files(url_to_target, num_threads=1):
     dl_items = list(url_to_target.items())
 
     with multiprocessing.pool.ThreadPool(num_threads) as p:
-        result = list(tqdm(
+        result = list(logger.progress(
             p.imap(_download_file, dl_items),
             total=len(dl_items),
-            desc='Download Files'
+            description='Download Files'
         ))
 
         return result
@@ -37,7 +42,7 @@ def _download_file(item):
     return download_file(item[0], item[1])
 
 
-def download_file(url, target_path, show_progress=False, num_threads=1):
+def download_file(url, target_path, num_threads=1):
     """
     Download the file from the given `url` and store it at `target_path`.
     Return a tuple x (url, bool, str).
@@ -50,7 +55,6 @@ def download_file(url, target_path, show_progress=False, num_threads=1):
         return download_file_parallel(
             url,
             target_path,
-            show_progress=show_progress,
             num_threads=num_threads
         )
 
@@ -62,25 +66,28 @@ def download_file(url, target_path, show_progress=False, num_threads=1):
             url
         ))
 
-    if show_progress:
-        file_size = int(requests.head(url).headers['Content-Length'])
-        pbar = tqdm(total=file_size, desc='Download File', unit_scale=True)
+    file_size = int(requests.head(url).headers['Content-Length'])
+    bytes_loaded = 0
+    bytes_since_last_log = 0
+    logger.info('Download file from "%s" with size: %d B', url, file_size)
 
     with open(target_path, 'wb') as f:
         for chunk in r.iter_content(chunk_size=1024):
             if chunk:
                 f.write(chunk)
+                bytes_loaded += 1024
+                bytes_since_last_log += 1024
 
-                if show_progress:
-                    pbar.update(1024)
+                if bytes_since_last_log >= PROGRESS_LOGGER_BYTE_DELAY:
+                    logger.info('Download [%06.2f%%]', bytes_loaded / file_size * 100)
+                    bytes_since_last_log = 0
 
-    if show_progress:
-        pbar.close()
+    logger.info('Finished download')
 
     return (url, True, target_path)
 
 
-def download_file_parallel(url, target_path, show_progress=False, num_threads=1):
+def download_file_parallel(url, target_path, num_threads=1):
     """
     Download the file from the given `url` and store it at `target_path`.
     Return a tuple x (url, bool, str).
@@ -92,24 +99,26 @@ def download_file_parallel(url, target_path, show_progress=False, num_threads=1)
     downloader = Downloader(url, target_path, num_threads)
     downloader.start()
 
-    if show_progress:
-        #
-        # Wait until we know file size
-        #
-        while downloader.total_length == 0:
-            pass
+    # Wait until we know file size
+    while downloader.total_length == 0:
+        pass
 
-        pbar = tqdm(total=downloader.total_length, desc='Download File', unit_scale=True)
+    file_size = downloader.total_length
+    logger.info('Download file from "%s" with size: %d B', url, file_size)
 
-        def update_pbar(x):
-            pbar.update(x.total_downloaded - pbar.n)
+    bytes_at_last_log = 0
 
-        downloader.subscribe(update_pbar, 10)
+    def callback(x):
+        nonlocal bytes_at_last_log
 
[…]
