+++ docs/reference/io.rst
@@ -44,7 +44,7 @@ Implementations
   Acoustic Event Dataset            x         x
   AudioMNIST                        x         x
   Broadcast                                   x
-  Common Voice                                x
+  Common Voice                      x         x
   Default                                     x      x
   ESC-50                            x         x
   Free-Spoken-Digit-Dataset         x         x
@@ -90,6 +90,9 @@ Broadcast
 
 Common-Voice
 ^^^^^^^^^^^^
+.. autoclass:: CommonVoiceDownloader
+   :members:
+
 .. autoclass:: CommonVoiceReader
    :members:
 
