+++ .deepsource.toml
@@ -1,22 +1,23 @@
 version = 1
 
 test_patterns = [
-  "tests/**"
+  "tests/**",
+  "bench/**"
 ]
 
 exclude_patterns = [
-  "bench/**",
   "examples/**",
   "docs/**"
 ]
 
-
 [[analyzers]]
 name = "python"
 enabled = true
-runtime_version = "3.x.x"
 
-# Test coverage analyzer
+  [analyzers.meta]
+  runtime_version = "3.x.x"
+  max_line_length = 120
+
 [[analyzers]]
 name = "test-coverage"
 enabled = true
