+++ audiomate/utils/audioread.py
@@ -10,4 +10,10 @@ available_backends = audioread.available_backends()
 
 
 def audio_open(path):
+    """
+    Just calls ``audioread.audio_open``,
+    but with backends cached in a global variable.
+    Brings better performance, since available backends
+    are evaluated only once.
+    """
     return audioread.audio_open(path, backends=available_backends)
