+++ audiomate/annotations/label_list.py
@@ -6,7 +6,7 @@ import intervaltree
 from .label import Label
 
 
-class LabelList(object):
+class LabelList:
     """
     Represents a list of labels which describe an utterance.
     An utterance can have multiple label-lists.
@@ -86,9 +86,10 @@ class LabelList(object):
     @property
     def total_length(self):
         """
-        Return the cumulative length of all labels. (Number of characters)
+        Return the cumulative length of all labels
+        (Number of characters).
         """
-        return sum([label.length for label in self.labels])
+        return sum(label.length for label in self.labels)
 
     #
     #   Alteration
@@ -105,9 +106,7 @@ class LabelList(object):
         self.label_tree.addi(label.start, label.end, label)
 
     def addl(self, value, start=0.0, end=float('inf')):
-        """
-        Shortcut for ``add(Label(value, start, end))``.
-        """
+        """ Shortcut for ``add(Label(value, start, end))``. """
         self.add(Label(value, start=start, end=end))
 
     def update(self, labels):
@@ -226,7 +225,8 @@ class LabelList(object):
 
     def label_total_duration(self):
         """
-        Return for each distinct label value the total duration of all occurrences.
+        Return for each distinct label value the total duration of
+        all occurrences.
 
         Returns:
             dict: A dictionary containing for every label-value (key)
@@ -270,7 +270,7 @@ class LabelList(object):
             ['a', 'b', 'c', 'd']
         """
 
-        all_labels = set([l.value for l in self])
+        all_labels = {l.value for l in self}
         return sorted(all_labels)
 
     def label_count(self):
@@ -305,8 +305,8 @@ class LabelList(object):
         Return a list of all tokens occurring in the label-list.
 
         Args:
-            delimiter (str): The delimiter used to split labels into tokens
-                             (see :meth:`audiomate.annotations.Label.tokenized`).
+            delimiter (str): The delimiter used to split labels into tokens.
+                             See :meth:`audiomate.annotations.Label.tokenized`
 
         Returns:
              :class:`set`: A set of distinct tokens.
@@ -326,12 +326,13 @@ class LabelList(object):
         """
         Return a string with all labels concatenated together.
         The order of the labels is defined by the start of the label.
-        If the overlapping between two labels is greater than ``overlap_threshold``,
-        an Exception is thrown.
+        If the overlapping between two labels is greater than
+        ``overlap_threshold``, an Exception is thrown.
 
         Args:
             delimiter (str): A string to join two consecutive labels.
-            overlap_threshold (float): Maximum overlap between two consecutive labels.
+            overlap_threshold (float): Maximum overlap between two
+                                       consecutive labels.
 
         Returns:
             str: A string with all labels concatenated together.
@@ -364,15 +365,18 @@ class LabelList(object):
         """
         Return a ordered list of tokens based on all labels.
         Joins all token from all labels (``label.tokenized()```).
-        If the overlapping between two labels is greater than ``overlap_threshold``,
-        an Exception is thrown.
+        If the overlapping between two labels is greater than
+        ``overlap_threshold``, an Exception is thrown.
 
         Args:
-            delimiter (str): The delimiter used to split labels into tokens. (default: space)
-            overlap_threshold (float): Maximum overlap between two consecutive labels.
+            delimiter (str): The delimiter used to split labels into tokens.
+                             (default: space)
+            overlap_threshold (float): Maximum overlap between two
+                                       consecutive labels.
 
         Returns:
-            str: A list containing tokens of all labels ordered according to the label order.
+            str: A list containing tokens of all labels ordered according
+                 to the label order.
 
         Example:
             >>> ll = LabelList(idx='some', labels=[
@@ -407,8 +411,8 @@ class LabelList(object):
         Create a separate Label-List for every distinct label-value.
 
         Returns:
[…]
