+++ shippable.yml
@@ -1,27 +1,0 @@
-language: python
-
-python:
-  - 3.5
-  - 3.6
-
-branches:
-  except:
-    - gh-pages
-
-install:
-  - sudo apt-get update
-  - sudo apt-get install -y libav-tools sox
-  - "pip install -e .[dev,ci]"
-
-before_script:
-  #Create folders for test and code coverage
-  - mkdir -p shippable/testresults
-  - mkdir -p shippable/codecoverage
-
-script:
-  # Run flake
-  - flake8
-
-  #Run test and code coverage and output results to the right folder
-  - pytest --junitxml=shippable/testresults/tests.xml
-  - pytest --cov=audiomate --cov-report=xml:shippable/codecoverage/coverage.xml
