### 2020-05-18 41604b9
add docs how to install sox using conda
