### 2020-05-18 1ac1799
add docs how to install sox using conda
