### 2020-05-18 5f3847a
Fix code sample
fixes #124
