### 2020-05-18 ynop/audiomate#125 @ynop
You are right.
I agree on both your suggestions:
- Fail if destination exists
- Optional flag to overwrite
