### 2020-05-18 ynop/audiomate#126 @ynop
Yes, that is good idea. I also had something similar in mind.
Due to some datasets changing "frequently", I wanted to introduce versions.
So you could actually select which version you want to use.
Of course this could be combined with your approach with tags and checksums.
