### 2020-05-18 eb0cf99
add conda sox install in rst
