### 2020-05-18 ynop/audiomate#126 @faroit
> Due to some datasets changing "frequently", I wanted to introduce versions.
So you could actually select which version you want to use.

yes, that's also a good idea. and could be added on top of tags. But I think you might end up with a less confusing API if one version of `audiomate` only supports a fixed amount of dataset versions. Of course, this would come with the drawback that users would be unable to load an older dataset version even though they use a new version of `audiomate`. But then I think this wouldn't happen in practice as most users would use audiomate for a single dataset per project.

For now, I would suggest to freeze as many downloads as possible instead of pointing to the latest.
