### 2020-05-18 #2135 @faroit
`audiomate` is a python based package that addresses the extraction, transformation and loading of audio datasets.
It comes with a large list of available datasets and many of them can be automatically downloaded from within python. The package is very well done and is excellently documented. It offers great value for the scientific audio community.

## Concerns
I added a few concerns in the issues opened by me. Most of them are minor can quickly be addressed or have already been addressed by the authors.

However, I have two main concerns that I would ask the authors to address before this paper gets accepted:

## Related software tools

The paper does not mention related software packages. In such section other tool would be listed and compared to `audiomate`.  Key differences would be highlighted as well as unique features  that emphasise the contributions in `audiomate`. Here are a few packages that are related:

### Audio related software
* https://github.com/juliagusak/dataloaders
* https://github.com/mir-dataset-loaders/mirdata
* https://github.com/mcfletch/audiodatasets
* https://pytorch.org/audio/datasets.html
* https://github.com/mpariente/asteroid

### Non-Audio related software
* https://pytorch.org/docs/stable/torchvision/datasets.html
* https://github.com/tensorflow/datasets

## Reproducibility

In [issue 126](https://github.com/ynop/audiomate/issues/126) I brought up reproducibility issued with audiomate. Compared to other software packages such as [mirdata](https://github.com/mir-dataset-loaders/mirdata) or [torchvision.datasets](https://pytorch.org/docs/stable/torchvision/datasets.html), `audiomate` does not come with mechanisms to allow users to reproduce dataset downloads. Often there are two ways to foster this:

* __verification__: making sure that a downloaded dataset matches a reference file. This is usually verified through the use of checksums/hashes in other packages.
* __versioning__: making sure that a downloaded dataset is the same when being downloaded at later times. This is usually archived by referring to specific tags (e.g. on github repos) or unique identifiers (e.g. DOIs on zenodo).

Both features are not covered by `audiomate` and I would highly suggest to incorporate at least __versioning__ into the software. E.g. this could easily be implemented by the authors by changing links from `master` branches of repos to specific commits/tags.

## Recommendation

For this paper being accepted in JOSS, I would suggest make these changes regarding a related work section in the paper and versioning the dataset downloads. However, I am happy to discuss this issues here with authors and other reviewers and I leave the final decision for @terrytangyuan
