### 2020-05-25 #2135 @ynop
Hi @faroit, thanks for your review.

Regarding reproducibility I completely agree with you, that versioning and verification would be cool. But this is a lot of work and I am not sure if we can do that in the near future. The only change that is possible with reasonable effort is using fixed commits for git-based datasets. The problem with that is that you have to update audiomate, every time a new version is released (and know that there is a new version). Furthermore, a user can fix the version of such datasets on her/his own by passing the specific url to the downloader.

Regarding related software, I will add a section in the paper.
