### 2020-05-25 #2135 @faroit
> The only change that is possible with reasonable effort is using fixed commits for git-based datasets.
> The problem with that is that you have to update audiomate, every time a new version is released (and know that there is a new version).

I think there is no way around this. Given how less frequent datasets are actually updated the workload is quite small, also I would say that reproducibility is more important for researchers than having the most recent version of the a dataset these . Then, this would currently only affect 3 datasets in audiomate, so this change would be rather quick, I guess.

> Furthermore, a user can fix the version of such datasets on her/his own by passing the specific url to the downloader.

Maybe have a look at [torch.hub](https://pytorch.org/docs/stable/hub.html#torch.hub.load) to implement a similar repo/tag scheme?
