### 2020-05-25 #2135 @ynop
Another thought: What happens If we update the version in audiomate. The user either has to keep working with the old audiomate version (and fix it) or use a custom download URL afterwards. I am not sure if it is a good approach to have the dataset versions depend on the software version, since this is not really transparent to the user.
