### 2020-05-25 31203e4
Add section about related work in joss paper
