+++ paper/paper.bib
@@ -60,3 +60,51 @@
     archivePrefix={arXiv},
     primaryClass={cs.CL}
 }
+
+@misc{torchaudio,
+    title = {TORCHAUDIO},
+    howpublished = {\url{https://pytorch.org/audio/}},
+    note = {Accessed: 2020-05-25},
+    year = {2020}
+}
+
+@misc{dataloaders,
+    title = {dataloaders},
+    howpublished = {\url{https://github.com/juliagusak/dataloaders}},
+    note = {Accessed: 2020-05-25},
+    year = {2020}
+}
+
+@misc{audiodatasets,
+    title = {Audio Datasets},
+    howpublished = {\url{https://github.com/mcfletch/audiodatasets}},
+    note = {Accessed: 2020-05-25},
+    year = {2020}
+}
+
+@misc{mirdata,
+    title = {mirdata},
+    howpublished = {\url{https://github.com/mir-dataset-loaders/mirdata}},
+    note = {Accessed: 2020-05-25},
+    year = {2020}
+}
+
+@misc{speechcorpusdownloader,
+    title = {Speech Corpus Downloader},
+    howpublished = {\url{https://github.com/mdangschat/speech-corpus-dl}},
+    note = {Accessed: 2020-05-25},
+    year = {2020}
+}
+
+@inproceedings {tensorflow,
+    author = {Mart{\'\i}n Abadi and Paul Barham and Jianmin Chen and Zhifeng Chen and Andy Davis and Jeffrey Dean and Matthieu Devin and Sanjay Ghemawat and Geoffrey Irving and Michael Isard and Manjunath Kudlur and Josh Levenberg and Rajat Monga and Sherry Moore and Derek G. Murray and Benoit Steiner and Paul Tucker and Vijay Vasudevan and Pete Warden and Martin Wicke and Yuan Yu and Xiaoqiang Zheng},
+    title = {TensorFlow: A System for Large-Scale Machine Learning},
+    booktitle = {12th {USENIX} Symposium on Operating Systems Design and Implementation ({OSDI} 16)},
+    year = {2016},
+    isbn = {978-1-931971-33-1},
+    address = {Savannah, GA},
+    pages = {265--283},
+    url = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi},
+    publisher = {{USENIX} Association},
+    month = nov,
+}
