### 2020-05-26 #2135 @faroit
@ynop

> what happens If we update the version in audiomate. The user either has to keep working with the old audiomate version (and fix it) or use a custom download URL afterwards. I am not sure if it is a good approach to have the dataset versions depend on the software version, since this is not really transparent to the user.

I did a bit of research and I found that most libraries actually do not support a good way to version the dataset. In many cases, the dataset is just verified with a checksum, so when a dataset is changed, the checksum fails.

A detailed solution is proposed by the [tensorflow tfds team](https://github.com/tensorflow/datasets/blob/master/docs/datasets_versioning.md). Basically they followed your idea to implement versions for each dataset. Now I agree that this is significant amount of work, so I agree to take this out of the JOSS review.
