### 2020-06-20 01ab691
Add progress logging bar.
