### 2020-06-20 a4b93a4
Add reader for css10 dataset.
