+++ audiomate/corpus/io/__init__.py
@@ -31,6 +31,8 @@ from .fluent_speech import FluentSpeechReader  # noqa: F401
 from .wav2letter import Wav2LetterWriter  # noqa: F401
 from .nvidia_jasper import NvidiaJasperWriter  # noqa: F401
 from .librispeech import LibriSpeechDownloader, LibriSpeechReader  # noqa: F401
+from .css_ten import CssTenReader  # noqa: F401
+
 
 __downloaders = {}
 for cls in CorpusDownloader.__subclasses__():
