### 2020-06-20 fe0e8c2
Add downloader and reader for zamia speech dataset.
