+++ audiomate/corpus/io/zamia_speech.py
@@ -1,0 +1,307 @@
+import os
+import re
+import shutil
+import tarfile
+
+import audiomate
+import requests
+from audiomate import annotations
+from audiomate import issuers
+from audiomate import logutil
+from audiomate.utils import download
+from audiomate.utils import textfile
+
+from . import base
+
+logger = logutil.getLogger()
+
+# ==================================================================================================
+
+urls_speech_dir = {
+    "de": "https://goofy.zamia.org/zamia-speech/corpora/zamia_de/",
+    "en": "https://goofy.zamia.org/zamia-speech/corpora/zamia_en/",
+}
+
+urls_transcripts = {
+    "de": "https://raw.githubusercontent.com/gooofy/zamia-speech/master/data/src/speech/zamia_de/transcripts_00.csv",
+    "en": "https://raw.githubusercontent.com/gooofy/zamia-speech/master/data/src/speech/zamia_en/transcripts_00.csv",
+}
+
+
+# ==================================================================================================
+
+
+class ZamiaSpeechDownloader(base.CorpusDownloader):
+    """ Most of the code is take from the Voxforge downloader as the structure is similar """
+
+    def __init__(self, lang="de"):
+        if lang in urls_speech_dir.keys():
+            self.url = urls_speech_dir[lang]
+        else:
+            msg = "There is no zamia-speech URL present for language {}!"
+            raise ValueError(msg.format(lang))
+
+    @classmethod
+    def type(cls):
+        return "zamia-speech"
+
+    # ==============================================================================================
+
+    def _download(self, target_path):
+        """ Download the data to target_path """
+
+        temp_folder = os.path.join(target_path, "download")
+        os.makedirs(temp_folder, exist_ok=True)
+
+        available = ZamiaSpeechDownloader.available_files(self.url)
+        logger.info("Found %d available archives to download", len(available))
+
+        downloaded = self.download_files(available, temp_folder)
+
+        logger.info("Extract %d files", len(downloaded))
+        ZamiaSpeechDownloader.extract_files(downloaded, target_path)
+
+        shutil.rmtree(temp_folder)
+
+    # ==============================================================================================
+
+    @staticmethod
+    def available_files(url):
+        """ Extract and return urls for all available .tgz files """
+
+        req = requests.get(url)
+        if req.status_code != 200:
+            msg = "Failed to download data (status {}) from {}!"
+            raise base.FailedDownloadException(msg.format(req.status_code, url))
+
+        page_content = req.text
+        link_pattern = re.compile(r'<a href="(.*?)">(.*?)</a>')
+        available_files = []
+
+        for match in link_pattern.findall(page_content):
+            if match[0].endswith(".tgz"):
+                available_files.append(os.path.join(url, match[0]))
+
+        return available_files
+
+    # ==============================================================================================
+
+    @staticmethod
+    def download_files(file_urls, target_path):
+        """ Download all files and store to the given path """
+
+        os.makedirs(target_path, exist_ok=True)
+        url_to_target = {}
+
+        for file_url in file_urls:
+            file_name = os.path.basename(file_url)
+            target_file_path = os.path.join(target_path, file_name)
+            url_to_target[file_url] = target_file_path
+
+        dl_result = download.download_files(url_to_target, num_threads=1)
+
+        downloaded_files = []
+        for url, status, path_or_msg in dl_result:
+            if status:
+                downloaded_files.append(path_or_msg)
+            else:
+                logger.info("Download failed for url %s", url)
+
+        return downloaded_files
+
+    # ==============================================================================================
+
+    @staticmethod
+    def extract_files(file_paths, target_path):
+        """ Unpack all files to the given path """
+
+        os.makedirs(target_path, exist_ok=True)
+        extracted = []
+
+        for file_path in file_paths:
[…]
