### 2020-06-20 755f85a
Exclude empty transcriptions. Better logging.
