### 2020-06-24 #2135 @arfon
> Now I agree that this is significant amount of work, so I agree to take this out of the JOSS review.

Thanks @faroit. Could you check on @ynop's most recent updates to see if you're able to check off the last few checkboxes for your review?
