### 2020-06-24 e5c08b0
Detect transcription file encoding type.
