+++ setup.py
@@ -21,7 +21,8 @@ REQUIRED = [
     'intervaltree == 3.0.2',
     'sox == 1.3.7',
     'PGet == 0.5.0',
-    'tqdm'
+    'tqdm',
+    'chardet'
 ]
 
 # Packages required for dev/ci enrionment
