### 2020-06-28 a142638
Add more languages to voxforge.
