### 2020-06-29 510cb34
Add spanish to cv links.
