### 2020-07-01 c4340c1
Fix numba version to prevent error in tests due to librosa
