+++ setup.py
@@ -21,6 +21,7 @@ REQUIRED = [
     'intervaltree == 3.0.2',
     'sox == 1.3.7',
     'PGet == 0.5.0',
+    'numba == 0.49.1'
 ]
 
 # Packages required for dev/ci enrionment
