### 2020-07-01 #2135 @faroit
> @faroit We added a section about related work.

Great. I made three comments on the changes:

* https://github.com/ynop/audiomate/commit/31203e4817be1626f3266aa5eab3cc6bcf811bd8#r40297165
* https://github.com/ynop/audiomate/commit/31203e4817be1626f3266aa5eab3cc6bcf811bd8#r40297291
* https://github.com/ynop/audiomate/commit/31203e4817be1626f3266aa5eab3cc6bcf811bd8#commitcomment-40297261

can you address these?
