### 2020-07-01 984a37c
Mention numpy in the joss paper
