### 2020-07-01 #2135 @ynop
@faroit We reference numpy in the second example now.
