### 2020-07-01 #2135 @faroit
> @faroit We reference numpy in the second example now.

Looks good
