### 2020-07-01 #2135 @arfon
@ynop - could you take a look at the suggested DOIs above from Whedon? If they're correct, then please add them to your BibTeX file with entries like this: `doi=10.1109/icassp.2015.7178964`
