### 2020-07-01 9084f9e
Update bibtex entry for librispeech
