+++ paper/paper.bib
@@ -25,7 +25,8 @@
     author={Vassil Panayotov and Guoguo Chen and Daniel Povey and Sanjeev Khudanpur},
     journal={2015 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
     year={2015},
-    pages={5206-5210}
+    pages={5206-5210},
+    doi=10.1109/icassp.2015.7178964
 }
 
 @article{deepspeech,
