### 2020-07-01 7c29c23
Fix paper bibliography
