+++ paper/paper.bib
@@ -26,7 +26,7 @@
     journal={2015 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
     year={2015},
     pages={5206-5210},
-    doi=10.1109/icassp.2015.7178964
+    doi={10.1109/icassp.2015.7178964}
 }
 
 @article{deepspeech,
