### 2020-07-01 #2135 @danielskatz
as you've probably learned, whedon only see one command at a time, and only if it's the first thing in a comment
