### 2020-07-01 ynop/audiomate#125 @cdkini
Is the issue specific to ESC50Downloader or should the change affect all classes that inherit from ArchiveDownloader?

I'd imagine the issue is prevelant for all such downloader classes but I want to make sure before changing the parent class.
