### 2020-07-01 ynop/audiomate#125 @ynop
This should apply for all downloaders, so you can implement directly in `CorpusDownloader`.
