### 2020-07-05 324df37
Added file existence check and option for force redownload to CorpusDownloader - GH #125
