### 2020-07-05 143c21d
Added file existence check and option for force redownload to CorpusDownloader - GH #125
