### 2020-07-06 def418c
Try two paths for CV audiofiles. Related to #4.
