### 2020-07-07 df9eda5
Add downloader and reader for lingualibre.
