+++ audiomate/corpus/io/__init__.py
@@ -33,6 +33,7 @@ from .nvidia_jasper import NvidiaJasperWriter  # noqa: F401
 from .librispeech import LibriSpeechDownloader, LibriSpeechReader  # noqa: F401
 from .css_ten import CssTenReader  # noqa: F401
 from .zamia_speech import ZamiaSpeechDownloader, ZamiaSpeechReader  # noqa: F401
+from .lingualibre import LinguaLibreDownloader, LinguaLibreReader  # noqa: F401
 
 
 __downloaders = {}
