+++ audiomate/corpus/io/lingualibre.py
@@ -1,0 +1,164 @@
+import hashlib
+import os
+
+import audiomate
+from audiomate import annotations
+from audiomate import issuers
+from audiomate import logutil
+from audiomate.utils import download
+
+from . import base
+
+logger = logutil.getLogger()
+
+# ==================================================================================================
+
+urls = {
+    "afr": "https://lingualibre.org/datasets/Q150-afr-Afrikaans.zip",
+    "amh": "https://lingualibre.org/datasets/Q154-amh-Amharic.zip",
+    "ara": "https://lingualibre.org/datasets/Q219-ara-Arabic.zip",
+    "arq": "https://lingualibre.org/datasets/Q6714-arq-AlgerianArabic.zip",
+    "ary": "https://lingualibre.org/datasets/Q264201-ary-MoroccanArabic.zip",
+    "atj": "https://lingualibre.org/datasets/Q52295-atj-Atikamekw.zip",
+    "bam": "https://lingualibre.org/datasets/Q318-bam-Bambara.zip",
+    "bas": "https://lingualibre.org/datasets/Q405-bas-Basaalanguage.zip",
+    "bbj": "https://lingualibre.org/datasets/Q52067-bbj-Ghomala%27%20language.zip",
+    "bci": "https://lingualibre.org/datasets/Q19858-bci-Baoul%C3%A9.zip",
+    "bcl": "https://lingualibre.org/datasets/Q115107-bcl-CentralBikol.zip",
+    "bdu": "https://lingualibre.org/datasets/Q52073-bdu-Oroko.zip",
+    "ben": "https://lingualibre.org/datasets/Q307-ben-Bengali.zip",
+    "bum": "https://lingualibre.org/datasets/Q52068-bum-Bululanguage.zip",
+    "bzm": "https://lingualibre.org/datasets/Q52074-bzm-Londo.zip",
+    "cat": "https://lingualibre.org/datasets/Q203-cat-Catalan.zip",
+    "ces": "https://lingualibre.org/datasets/Q392-ces-Czech.zip",
+    "cmn": "https://lingualibre.org/datasets/Q113-cmn-MandarinChinese.zip",
+    "cym": "https://lingualibre.org/datasets/Q141-cym-Welsh.zip",
+    "deu": "https://lingualibre.org/datasets/Q24-deu-German.zip",
+    "dua": "https://lingualibre.org/datasets/Q52071-dua-Duala.zip",
+    "dyu": "https://lingualibre.org/datasets/Q159-dyu-Dioulalanguage.zip",
+    "eng": "https://lingualibre.org/datasets/Q22-eng-English.zip",
+    "epo": "https://lingualibre.org/datasets/Q25-epo-Esperanto.zip",
+    "eus": "https://lingualibre.org/datasets/Q299-eus-Basque.zip",
+    "fin": "https://lingualibre.org/datasets/Q33-fin-Finnish.zip",
+    "fon": "https://lingualibre.org/datasets/Q242-fon-Fon.zip",
+    "fra": "https://lingualibre.org/datasets/Q21-fra-French.zip",
+    "gaa": "https://lingualibre.org/datasets/Q321-gaa-Ga.zip",
+    "gcf": "https://lingualibre.org/datasets/Q83641-gcf-GuadeloupeanCreoleFrench.zip",
+    "gre": "https://lingualibre.org/datasets/Q205-gre-Greek.zip",
+    "hat": "https://lingualibre.org/datasets/Q165-hat-HaitianCreole.zip",
+    "hav": "https://lingualibre.org/datasets/Q51299-hav-Havu.zip",
+    "heb": "https://lingualibre.org/datasets/Q397-heb-Hebrew.zip",
[…]
