+++ audiomate/corpus/io/common_voice.py
@@ -1,74 +1,79 @@
-import os
 import glob
+import os
 
 import audiomate
+import pandas as pd
 from audiomate import annotations
 from audiomate import issuers
 from audiomate import logutil
 from audiomate.corpus import subset
-from audiomate.utils import textfile
+
 from . import base
 from . import downloader
 
 logger = logutil.getLogger()
 
-# DOWNLOAD_URLS taken from https://voice.mozilla.org/de/datasets
+# ==================================================================================================
+
 DOWNLOAD_URLS = {
-    'ab': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/ab.tar.gz',  # noqa
-    'ar': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/ar.tar.gz',  # noqa
-    'as': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/as.tar.gz',  # noqa
-    'br': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/br.tar.gz',  # noqa
-    'ca': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/ca.tar.gz',  # noqa
-    'cnh': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/cnh.tar.gz',  # noqa
-    'cs': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/cs.tar.gz',  # noqa
-    'cv': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/cv.tar.gz',  # noqa
-    'cy': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/cy.tar.gz',  # noqa
-    'de': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/de.tar.gz',  # noqa
-    'dv': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/dv.tar.gz',  # noqa
-    'el': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/el.tar.gz',  # noqa
-    'en': 'https://voice-prod-bundler-ee1969a6ce8178826482b88e843c335139bd3fb4.s3.amazonaws.com/cv-corpus-5-2020-06-22/en.tar.gz',  # noqa
[…]
