### 2020-07-08 fa2faf3
Improve language link handling.
