### 2020-07-08 a3c82f4
Fix logger printing.
