### 2020-07-08 febb975
Fix mailabs variable subdirectories. Closes #4.
