+++ audiomate/corpus/io/mailabs.py
@@ -3,60 +3,56 @@ import os
 import audiomate
 from audiomate import annotations
 from audiomate import issuers
+from audiomate import logutil
 from audiomate.corpus import subset
-from audiomate.utils import download
 from audiomate.utils import textfile
 
 from . import base
+from . import downloader
 
-DOWNLOAD_URLS = {
-    'de_DE': 'https://www.caito.de/data/Training/stt_tts/de_DE.tgz',
-    'en_UK': 'https://www.caito.de/data/Training/stt_tts/en_UK.tgz',
-    'en_US': 'https://www.caito.de/data/Training/stt_tts/en_US.tgz',
-    'es_ES': 'https://www.caito.de/data/Training/stt_tts/es_ES.tgz',
-    'it_IT': 'https://www.caito.de/data/Training/stt_tts/it_IT.tgz',
-    'uk_UK': 'https://www.caito.de/data/Training/stt_tts/uk_UK.tgz',
-    'ru_RU': 'https://www.caito.de/data/Training/stt_tts/ru_RU.tgz',
-    'fr_FR': 'https://www.caito.de/data/Training/stt_tts/fr_FR.tgz',
-    'pl_PL': 'https://www.caito.de/data/Training/stt_tts/pl_PL.tgz',
+logger = logutil.getLogger()
+
+# ==================================================================================================
+
+BASE_URL = "https://www.caito.de/data/Training/stt_tts/{}.tgz"
+LANGUAGES = {
+    "de_DE",
+    "en_UK",
+    "en_US",
+    "es_ES",
+    "fr_FR",
+    "it_IT",
+    "pl_PL",
+    "ru_RU",
+    "uk_UK",
 }
 
 
-class MailabsDownloader(base.CorpusDownloader):
+# ==================================================================================================
+
+
+class MailabsDownloader(downloader.ArchiveDownloader):
     """
     Downloader for the M-AILABS Speech Dataset.
 
-    Args:
-        tags (list): List of tags for different parts to download.
-                     Corresponds to the tags in the
-                     `Statistics & Download Links` on the webpage.
-                     If ``None``, all parts are downloaded.
-        num_threads (int): Number of threads to use for download files.
     """
 
-    def __init__(self, tags=None, num_threads=1):
-        self.tags = tags
-        self.num_threads = num_threads
+    def __init__(self, lang="de_DE"):
+        if lang in LANGUAGES:
+            link = BASE_URL.format(lang)
+            super(MailabsDownloader, self).__init__(
+                link, move_files_up=True, num_threads=1
+            )
+        else:
+            msg = "There is no mailabs URL present for language {}!"
+            raise ValueError(msg.format(lang))
 
     @classmethod
     def type(cls):
-        return 'mailabs'
+        return "mailabs"
 
-    def _download(self, target_path):
-        os.makedirs(target_path, exist_ok=True)
 
-        for tag, download_url in DOWNLOAD_URLS.items():
-            if self.tags is None or tag in self.tags:
-                tmp_file = os.path.join(target_path, 'tmp_{}.tgz'.format(tag))
-
-                download.download_file(
-                    download_url,
-                    tmp_file,
-                    num_threads=self.num_threads
-                )
-                download.extract_tar(tmp_file, target_path)
-
-                os.remove(tmp_file)
+# ==================================================================================================
 
 
 class MailabsReader(base.CorpusReader):
@@ -71,26 +67,39 @@ class MailabsReader(base.CorpusReader):
 
     @classmethod
     def type(cls):
-        return 'mailabs'
+        return "mailabs"
 
     def _check_for_missing_files(self, path):
         return []
 
+    # ==============================================================================================
+
     def _load(self, path):
[…]
