### 2020-07-12 5da5e70
Added file existence check to CorpusReader ABC - Fixes #89
