### 2020-07-12 4bfca59
Added file existence check to CorpusReader ABC - Fixes #89
