### 2020-08-03 59561d2
Bump version: 5.2.0 → 6.0.0
