+++ docs/conf.py
@@ -70,9 +70,9 @@ author = 'Matthias Büchi, Andreas Ahlenstorf'
 # built documents.
 #
 # The short X.Y version.
-version = '5.2.0'
+version = '6.0.0'
 # The full version, including alpha/beta/rc tags.
-release = '5.2.0'
+release = '6.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
