+++ setup.py
@@ -69,7 +69,7 @@ except FileNotFoundError:
 ##################################################
 
 setup(name='audiomate',
-      version='5.2.0',
+      version='6.0.0',
       description=DESCRIPTION,
       long_description=long_description,
       long_description_content_type='text/markdown',
