+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 5.2.0
+current_version = 6.0.0
 commit = True
 tag = True
 parse = (?P<major>\d+)
@@ -24,4 +24,3 @@ replace = version='{new_version}'
 [bumpversion:file:docs/conf.py]
 
 [bumpversion:file:audiomate/__init__.py]
-
