### 2020-08-03 #2135 @ynop
@terrytangyuan New version is `v6.0.0` and DOI is `10.5281/zenodo.3970567`
