### 2020-08-03 #2135 @danielskatz
@openjournals/dev - note that in the XML there are a bunch of badly formed URLs in unstructured citations.  This is a bug we've had for a while.
1. What should we do in this case?
2. Can we fix the bug?
