+++ paper/paper.bib
@@ -100,7 +100,7 @@
 @inproceedings {tensorflow,
     author = {Mart{\'\i}n Abadi and Paul Barham and Jianmin Chen and Zhifeng Chen and Andy Davis and Jeffrey Dean and Matthieu Devin and Sanjay Ghemawat and Geoffrey Irving and Michael Isard and Manjunath Kudlur and Josh Levenberg and Rajat Monga and Sherry Moore and Derek G. Murray and Benoit Steiner and Paul Tucker and Vijay Vasudevan and Pete Warden and Martin Wicke and Yuan Yu and Xiaoqiang Zheng},
     title = {TensorFlow: A System for Large-Scale Machine Learning},
-    booktitle = {12th {USENIX} Symposium on Operating Systems Design and Implementation ({OSDI} 16)},
+    booktitle = {{12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)}},
     year = {2016},
     isbn = {978-1-931971-33-1},
     address = {Savannah, GA},
